+++ clikt/src/jvmMain/kotlin/com/github/ajalt/clikt/sources/PropertiesValueSource.kt
@@ -10,7 +10,6 @@ import java.util.*
 /**
  * A [ValueSource] that reads values from a [Properties] object.
  */
-@ExperimentalValueSourceApi
 object PropertiesValueSource {
     /**
      * Parse a properties [file] into a value source.