+++ test/src/jvmTest/kotlin/com/github/ajalt/clikt/sources/PropertiesValueSourceTest.kt
@@ -20,7 +20,6 @@ import org.junit.Test
 import org.junit.rules.TemporaryFolder
 import java.io.File
 
-@OptIn(ExperimentalValueSourceApi::class)
 class PropertiesValueSourceTest {
     @get:Rule
     var testFolder = TemporaryFolder()